Add files via upload
In recent days hacking of data has taken place more in the cloud. The data in public is stored confidentially but it hacks by the end users may be a hacker. So, the goal of this paper is to build an Android application that enables users to store confidential data in the cloud even though if the cloud is hacked, the hacker can only see the image but not the data behind the photo. Pictures are ideal for data hiding after hiding by the image it occupies a large amount of redundant space. Since this ought to be conceivable in additional ways than one, IS has been used to show for the data hiding the LSB technique in which the Most un-Immense Digit of every single byte is changed to shape the cycle string tending to the report behind the image. This methodology works honourably for 24-cycle record assortment pictures, it has not been as productive when using an 8-bit file assortment picture, use of the MSB technique where-using bit differencing—the hidden hdatacanberetrievedonlyby crypto techniques.